Backend/django/requirements.txt
django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-keycloak==5.3.1
boto3==1.38.0
dj-database-url==2.3.0
python-dotenv==1.1.0
requests==2.32.3
